+++ src/client/src/Interface/DLGs/EventDialog.cpp
@@ -323,7 +323,7 @@ bool CEventDialog::Create(const char* IDD) {
       CWinCtrl* pCtrl = Find( IID_BG_IMAGE );
       if ( nullptr != pCtrl && CTRL_IMAGE == pCtrl->GetControlType() ) {
         CTImage* pImage = (CTImage*)pCtrl;
-        pImage->SetScaleWidth( g_pCApp->GetWIDTH() / pImage->GetWidth() );
+        pImage->SetScaleWidth( (float)(g_pCApp->GetWIDTH() / pImage->GetWidth()) );
       }
 
       pCtrl = Find( IID_BTN_CLOSE );
